refactor: git conflict 해결
