+++ android/app/src/main/java/com/zzang/chongdae/presentation/view/home/HomeFragment.kt
@@ -7,12 +7,13 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
+import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.zzang.chongdae.data.remote.api.NetworkManager
-import com.zzang.chongdae.data.remote.source.impl.OfferingsDataSourceImpl
-import com.zzang.chongdae.data.repository.remote.OfferingsRepositoryImpl
+import com.zzang.chongdae.ChongdaeApp
+import com.zzang.chongdae.R
 import com.zzang.chongdae.databinding.FragmentHomeBinding
+import com.zzang.chongdae.presentation.view.MainActivity
 import com.zzang.chongdae.presentation.view.home.adapter.OfferingAdapter
 import com.zzang.chongdae.presentation.view.offeringdetail.OfferingDetailActivity
 
@@ -25,7 +26,7 @@ class HomeFragment : Fragment(), OnOfferingClickListener {
             offeringRepository = (requireActivity().application as ChongdaeApp).offeringRepository,
         )
     }
-    
+
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
@@ -34,7 +35,7 @@ class HomeFragment : Fragment(), OnOfferingClickListener {
         initBinding(inflater, container)
         return binding.root
     }
-    
+
     override fun onViewCreated(
         view: View,
         savedInstanceState: Bundle?,
@@ -50,7 +51,7 @@ class HomeFragment : Fragment(), OnOfferingClickListener {
 
         offeringAdapter.refresh()
     }
-    
+
     override fun onResume() {
         super.onResume()
         (activity as MainActivity).showBottomNavigation()
@@ -89,7 +90,7 @@ class HomeFragment : Fragment(), OnOfferingClickListener {
     override fun onClick(offeringId: Long) {
         OfferingDetailActivity.startActivity(activity as Context, offeringId)
     }
-    
+
     private fun navigateToOfferingWriteFragment() {
         binding.fabCreateOffering.setOnClickListener {
             findNavController().navigate(R.id.action_home_fragment_to_offering_write_fragment)
